fix: responds with 404 when post with requested ID is not found (fixes #25)

+++ src/main/java/dev/neubert/backendsystems/socialmedia/adapters/in/api/controllers/PostWebController.java
@@ -99,6 +99,9 @@ public class PostWebController {
             PostDto model
     ) {
         setCacheControlFiveMinutes();
+        if (this.postAdapter.getPostById(id) == null) {
+            return Response.status(HttpResponseStatus.NOT_FOUND.code()).build();
+        }
         var result = this.postAdapter.updatePost(id, model);
         return Response.status(HttpResponseStatus.NO_CONTENT.code())
                        .header("Location", createLocationHeader(result))
@@ -116,6 +119,9 @@ public class PostWebController {
             long id
     ) {
         var toBeDeleted = this.postAdapter.getPostById(id);
+        if (toBeDeleted == null) {
+            return Response.status(HttpResponseStatus.NOT_FOUND.code()).build();
+        }
         if (this.postAdapter.deletePost(toBeDeleted)) {
             return Response.status(HttpResponseStatus.NO_CONTENT.code()).build();
         } else {
